I'm going to murder something

+++ src/main/java/eu/crushedpixel/camerastudio/api/paths/CameraPosition.java
@@ -12,9 +12,9 @@ import java.util.UUID;
 public class CameraPosition {
 
 	private UUID world;
-	private int x;
-	private int y;
-	private int z;
+	private double x;
+	private double y;
+	private double z;
 	private float pitch;
 	private float yaw;
